+++ back/src/main/java/com/ll/goohaeyou/application/application/ApplicationService.java
@@ -11,7 +11,7 @@ import com.ll.goohaeyou.member.member.domain.Member;
 import com.ll.goohaeyou.member.member.application.MemberService;
 import com.ll.goohaeyou.global.event.notification.ApplicationCreateAndChangedEvent;
 import com.ll.goohaeyou.application.exception.ApplicationException;
-import com.ll.goohaeyou.global.exception.auth.AuthException;
+import com.ll.goohaeyou.auth.exception.AuthException;
 import com.ll.goohaeyou.jobPost.jobPost.exception.JobPostException;
 import lombok.RequiredArgsConstructor;
 import org.springframework.context.ApplicationEventPublisher;
