+++ back/src/main/java/com/ll/goohaeyou/application/prsentation/ApplicationController.java
@@ -4,7 +4,7 @@ import com.ll.goohaeyou.application.dto.ApplicationDto;
 import com.ll.goohaeyou.application.dto.ApplicationForm;
 import com.ll.goohaeyou.application.application.ApplicationService;
 import com.ll.goohaeyou.global.apiResponse.ApiResponse;
-import com.ll.goohaeyou.global.security.MemberDetails;
+import com.ll.goohaeyou.auth.domain.MemberDetails;
 import com.ll.goohaeyou.global.standard.base.Empty;
 import io.swagger.v3.oas.annotations.Operation;
 import io.swagger.v3.oas.annotations.tags.Tag;
