+++ back/src/main/java/com/ll/goohaeyou/auth/application/AuthenticationService.java
@@ -1,26 +1,21 @@
-package com.ll.goohaeyou.global.security.OAuth;
+package com.ll.goohaeyou.auth.application;
 
-import com.ll.goohaeyou.member.member.domain.Member;
-import com.ll.goohaeyou.member.member.domain.RefreshToken;
-import com.ll.goohaeyou.member.member.domain.repository.RefreshTokenRepository;
+import com.ll.goohaeyou.auth.domain.RefreshToken;
+import com.ll.goohaeyou.auth.domain.RefreshTokenRepository;
 import com.ll.goohaeyou.member.member.application.MemberService;
-import com.ll.goohaeyou.global.security.JwtTokenProvider;
+import com.ll.goohaeyou.member.member.dto.MemberDto;
+import com.ll.goohaeyou.member.member.domain.Member;
 import com.ll.goohaeyou.global.standard.base.util.CookieUtil;
 import jakarta.servlet.http.HttpServletRequest;
 import jakarta.servlet.http.HttpServletResponse;
 import lombok.RequiredArgsConstructor;
-import org.springframework.beans.factory.annotation.Value;
-import org.springframework.security.core.Authentication;
-import org.springframework.security.oauth2.core.user.OAuth2User;
-import org.springframework.security.web.authentication.SimpleUrlAuthenticationSuccessHandler;
-import org.springframework.stereotype.Component;
+import org.springframework.stereotype.Service;
 
-import java.io.IOException;
 import java.time.Duration;
 
+@Service
 @RequiredArgsConstructor
-@Component
-public class OAuth2SuccessHandler extends SimpleUrlAuthenticationSuccessHandler {
+public class AuthenticationService {
     public static final String REFRESH_TOKEN_COOKIE_NAME = "refresh_token";
     public static final String ACCESS_TOKEN_COOKIE_NAME = "access_token";
     public static final Duration REFRESH_TOKEN_DURATION = Duration.ofDays(1);
@@ -28,16 +23,10 @@ public class OAuth2SuccessHandler extends SimpleUrlAuthenticationSuccessHandler 
 
     private final JwtTokenProvider jwtTokenProvider;
     private final RefreshTokenRepository refreshTokenRepository;
-    private final OAuth2AuthorizationRequestBasedOnCookieRepository authorizationRequestRepository;
     private final MemberService memberService;
 
-    @Value("${custom.site.frontUrl}/member/socialLoginCallback")
-    private String redirectPath;
-
-    @Override
-    public void onAuthenticationSuccess(HttpServletRequest request, HttpServletResponse response, Authentication authentication) throws IOException {
-        OAuth2User oAuth2User = (OAuth2User) authentication.getPrincipal();
-        Member member = memberService.getMember((String) oAuth2User.getAttributes().get("email"));
+    public MemberDto authenticateAndSetTokens(String username, HttpServletRequest request, HttpServletResponse response) {
+        Member member = memberService.getMember(username);
 
         // 리프레쉬 토큰
         String refreshToken = jwtTokenProvider.generateToken(member, REFRESH_TOKEN_DURATION);
@@ -48,11 +37,10 @@ public class OAuth2SuccessHandler extends SimpleUrlAuthenticationSuccessHandler 
         String accessToken = jwtTokenProvider.generateToken(member, ACCESS_TOKEN_DURATION);
         addTokenToCookie(request, response, ACCESS_TOKEN_COOKIE_NAME, accessToken, ACCESS_TOKEN_DURATION);
 
-        clearAuthenticationAttributes(request, response);
-        getRedirectStrategy().sendRedirect(request, response, redirectPath);
+        return MemberDto.from(member);
     }
 
-    // 리프레쉬 토큰 DB 저장
+    // 리프레쉬 토큰을 DB에 저장
     private void saveRefreshToken(Long userId, String newRefreshToken) {
         RefreshToken refreshToken = refreshTokenRepository.findByUserId(userId)
                 .map(entity -> entity.update(newRefreshToken))
@@ -61,16 +49,10 @@ public class OAuth2SuccessHandler extends SimpleUrlAuthenticationSuccessHandler 
         refreshTokenRepository.save(refreshToken);
     }
 
-    // 토큰 쿠키 저장
+    // 토큰을 쿠키에 저장
     private void addTokenToCookie(HttpServletRequest request, HttpServletResponse response, String cookieName, String token, Duration duration) {
         int cookieMaxAge = (int) duration.toSeconds();
         CookieUtil.deleteCookie(request, response, cookieName);
         CookieUtil.addCookie(response, cookieName, token, cookieMaxAge);
-    }
-
-    //인증 과정 중 생긴 요청 속성 제거
-    private void clearAuthenticationAttributes(HttpServletRequest request, HttpServletResponse response) {
-        super.clearAuthenticationAttributes(request);
-        authorizationRequestRepository.removeAuthorizationRequestCookies(request, response);
     }
 }
