+++ back/src/main/java/com/ll/goohaeyou/auth/application/OAuth2UserCustomService.java
@@ -1,4 +1,4 @@
-package com.ll.goohaeyou.global.security.OAuth;
+package com.ll.goohaeyou.auth.application;
 
 import com.ll.goohaeyou.member.member.domain.Member;
 import com.ll.goohaeyou.member.member.domain.repository.MemberRepository;
