+++ back/src/main/java/com/ll/goohaeyou/auth/domain/MemberDetails.java
@@ -1,8 +1,8 @@
-package com.ll.goohaeyou.global.security;
+package com.ll.goohaeyou.auth.domain;
 
+import com.ll.goohaeyou.member.member.domain.type.Role;
 import com.ll.goohaeyou.member.member.domain.Member;
 import com.ll.goohaeyou.member.member.domain.type.Gender;
-import com.ll.goohaeyou.member.member.domain.type.Role;
 import lombok.RequiredArgsConstructor;
 import org.springframework.security.core.GrantedAuthority;
 import org.springframework.security.core.userdetails.UserDetails;
