+++ back/src/main/java/com/ll/goohaeyou/auth/dto/CreateAccessTokenRequest.java
@@ -1,4 +1,4 @@
-package com.ll.goohaeyou.member.member.dto;
+package com.ll.goohaeyou.auth.dto;
 
 import lombok.Getter;
 import lombok.Setter;
