+++ back/src/main/java/com/ll/goohaeyou/chat/message/presentation/MessageController.java
@@ -5,7 +5,7 @@ import com.ll.goohaeyou.chat.message.dto.MessageForm;
 import com.ll.goohaeyou.chat.message.domain.Message;
 import com.ll.goohaeyou.chat.message.application.MessageService;
 import com.ll.goohaeyou.global.apiResponse.ApiResponse;
-import com.ll.goohaeyou.global.security.MemberDetails;
+import com.ll.goohaeyou.auth.domain.MemberDetails;
 import com.ll.goohaeyou.global.standard.base.Empty;
 import io.swagger.v3.oas.annotations.Operation;
 import io.swagger.v3.oas.annotations.tags.Tag;
