+++ back/src/main/java/com/ll/goohaeyou/jobPost/employ/presentation/WorkCompletionController.java
@@ -2,7 +2,7 @@ package com.ll.goohaeyou.jobPost.employ.presentation;
 
 import com.ll.goohaeyou.jobPost.employ.application.WorkCompletionService;
 import com.ll.goohaeyou.global.apiResponse.ApiResponse;
-import com.ll.goohaeyou.global.security.MemberDetails;
+import com.ll.goohaeyou.auth.domain.MemberDetails;
 import com.ll.goohaeyou.global.standard.base.Empty;
 import io.swagger.v3.oas.annotations.Operation;
 import io.swagger.v3.oas.annotations.tags.Tag;
