+++ back/src/main/java/com/ll/goohaeyou/jobPost/jobPost/presentation/InterestController.java
@@ -3,7 +3,7 @@ package com.ll.goohaeyou.jobPost.jobPost.presentation;
 import com.ll.goohaeyou.jobPost.jobPost.application.InterestService;
 import com.ll.goohaeyou.jobPost.jobPost.application.JobPostService;
 import com.ll.goohaeyou.global.apiResponse.ApiResponse;
-import com.ll.goohaeyou.global.security.MemberDetails;
+import com.ll.goohaeyou.auth.domain.MemberDetails;
 import com.ll.goohaeyou.global.standard.base.Empty;
 import io.swagger.v3.oas.annotations.Operation;
 import io.swagger.v3.oas.annotations.tags.Tag;
