+++ back/src/main/java/com/ll/goohaeyou/member/emailAuth/presentation/EmailAuthController.java
@@ -3,7 +3,7 @@ package com.ll.goohaeyou.member.emailAuth.presentation;
 
 import com.ll.goohaeyou.member.emailAuth.application.EmailAuthService;
 import com.ll.goohaeyou.global.apiResponse.ApiResponse;
-import com.ll.goohaeyou.global.security.MemberDetails;
+import com.ll.goohaeyou.auth.domain.MemberDetails;
 import com.ll.goohaeyou.global.standard.base.Empty;
 import io.swagger.v3.oas.annotations.Operation;
 import io.swagger.v3.oas.annotations.tags.Tag;
