+++ back/src/main/java/com/ll/goohaeyou/member/member/presentation/MemberController.java
@@ -4,11 +4,11 @@ import com.ll.goohaeyou.member.member.dto.JoinForm;
 import com.ll.goohaeyou.member.member.dto.LoginForm;
 import com.ll.goohaeyou.member.member.dto.MemberDto;
 import com.ll.goohaeyou.member.member.dto.SocialProfileForm;
-import com.ll.goohaeyou.member.member.application.AuthenticationService;
+import com.ll.goohaeyou.auth.application.AuthenticationService;
 import com.ll.goohaeyou.member.member.application.MemberService;
 import com.ll.goohaeyou.notification.application.NotificationService;
 import com.ll.goohaeyou.global.apiResponse.ApiResponse;
-import com.ll.goohaeyou.global.security.MemberDetails;
+import com.ll.goohaeyou.auth.domain.MemberDetails;
 import com.ll.goohaeyou.global.standard.base.util.CookieUtil;
 import io.swagger.v3.oas.annotations.Operation;
 import io.swagger.v3.oas.annotations.tags.Tag;
