+++ back/src/main/java/com/ll/goohaeyou/payment/payment/service/PaymentInfoService.java
@@ -3,7 +3,7 @@ package com.ll.goohaeyou.payment.payment.service;
 import com.ll.goohaeyou.payment.payment.dto.PaymentDto;
 import com.ll.goohaeyou.payment.payment.domain.Payment;
 import com.ll.goohaeyou.payment.payment.domain.repository.PaymentRepository;
-import com.ll.goohaeyou.global.exception.auth.AuthException;
+import com.ll.goohaeyou.auth.exception.AuthException;
 import com.ll.goohaeyou.payment.payment.exception.PaymentException;
 import lombok.RequiredArgsConstructor;
 import org.springframework.stereotype.Service;
